requirements.txt
langchain==0.0.208
deeplake
openai
tiktoken
newspaper3k
python-dotenv
streamlit
streamlit_chat